Add command to report on build requests
