+++ java-components/adhoc-analyser-tool/src/main/java/com/redhat/hacbs/analyser/EntryPoint.java
@@ -4,6 +4,7 @@ import com.redhat.hacbs.analyser.artifactanalysis.AnalyseRepositoriesCommand;
 import com.redhat.hacbs.analyser.artifactanalysis.CheckoutRepositoriesCommand;
 import com.redhat.hacbs.analyser.artifactanalysis.ConsolidateScmInfoCommand;
 import com.redhat.hacbs.analyser.github.GithubCommand;
+import com.redhat.hacbs.analyser.kube.ListBuildRequestsCommand;
 import com.redhat.hacbs.analyser.pnc.PncCommand;
 
 import io.quarkus.picocli.runtime.annotations.TopCommand;
@@ -11,7 +12,7 @@ import picocli.CommandLine;
 
 @TopCommand
 @CommandLine.Command(mixinStandardHelpOptions = true, subcommands = { PncCommand.class, AnalyseRepositoriesCommand.class,
-        ConsolidateScmInfoCommand.class, GithubCommand.class, CheckoutRepositoriesCommand.class
+        ConsolidateScmInfoCommand.class, GithubCommand.class, CheckoutRepositoriesCommand.class, ListBuildRequestsCommand.class
 })
 public class EntryPoint {
 }
