+++ java-components/resource-model/src/main/java/com/redhat/hacbs/resources/model/v1alpha1/ArtifactBuildRequestStatus.java
@@ -1,44 +1,17 @@
 package com.redhat.hacbs.resources.model.v1alpha1;
 
 public class ArtifactBuildRequestStatus {
-    public enum State {
-        /**
-         * A new resource that has not been acted on by the operator
-         */
-        NEW,
-        /**
-         * The discovery pipeline is running to try and figure out how to build this artifact
-         */
-        DISCOVERING,
-        /**
-         * The discovery pipeline failed to find a way to build this
-         */
-        MISSING,
-        /**
-         * The build is running
-         */
-        BUILDING,
-        /**
-         * The build failed
-         */
-        FAILED,
-        /**
-         * The build completed successfully, the resource can be removed
-         */
-        COMPLETE
 
-    }
-
-    private State state = State.NEW;
+    private String state = "";
     private String recipeGitHash; //todo multiple git repos
     private String buildPipelineName; //todo multiple git repos
     private String message;
 
-    public State getState() {
+    public String getState() {
         return state;
     }
 
-    public ArtifactBuildRequestStatus setState(State state) {
+    public ArtifactBuildRequestStatus setState(String state) {
         this.state = state;
         return this;
     }
